refactor parser dir to parse
